test/requirements-test.txt
geopandas
rtree
pytest-qgis
pytest-qt
pytest-cov
pytest-mock